Use Storage on photos and avatars

+++ app/Actions/Jetstream/DeleteTeam.php
@@ -14,14 +14,15 @@ class DeleteTeam implements DeletesTeams
     public function delete(Team $team): void
     {
         // -----------------------------------------------------------------------
-        // delete team photo folder
+        // delete team photos and avatars folders
         // -----------------------------------------------------------------------
-        Storage::disk('photos')->deleteDirectory($team->id);
-        
-        // -----------------------------------------------------------------------
-        // delete team avatars folder
-        // -----------------------------------------------------------------------
-        Storage::disk('avatars')->deleteDirectory($team->id);
+        if (storage::disk('photos')->exists($team->id)) {
+            Storage::disk('photos')->deleteDirectory($team->id);
+        }
+
+        if (storage::disk('avatars')->exists($team->id)) {
+            Storage::disk('avatars')->deleteDirectory($team->id);
+        }
         // -----------------------------------------------------------------------
 
         $team->purge();
